test(nightwatch): remove login via login.zeit case and add conftest.py
Refs: ENG-160
