+++ tests/test_basics.py
@@ -2,11 +2,6 @@ import pytest
 
 
 @pytest.fixture(scope="session")
-def nightwatch_config():
-    return {"browser": {"baseurl": "https://httpbin.org"}}
-
-
-@pytest.fixture(scope="session")
 def base_url(nightwatch_config):
     return nightwatch_config.get("browser", {}).get("baseurl")
